Add files via upload
Tis is the outline that we discussed yesterday. I believe that an essential part of the new AI world of Quantum Computing depends on how closely linked "Language" is to QM and computing in general.  this a project on which i have been working on alone for many years.  As for brief background, other than being a AI researcher for many years (my first job was with Bell Labs in Holmdel, NJ in 70's), I was also a simultaneous interpreter for the US government (State Dept., Justice Dept., "Other"